requirements.txt
azure.mgmt.resource==21.2.1
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-mgmt-extendedlocation
azure.mgmt.network
pyfiglet
colorama



